src/main/java/guru/springframework/spring5recipeapp/controllers/IndexController.java
package guru.springframework.spring5recipeapp.controllers;

import guru.springframework.spring5recipeapp.model.Category;
import guru.springframework.spring5recipeapp.model.UnitOfMeasure;
import guru.springframework.spring5recipeapp.repositories.CategoryRepository;
import guru.springframework.spring5recipeapp.repositories.UnitOfMeasureRepository;
import guru.springframework.spring5recipeapp.service.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.Optional;

@Controller
@Slf4j
public class IndexController {

    private final RecipeService recipeService;

    public IndexController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @RequestMapping({"", "/", "/indexp"})
    public String getIndexPage(Model model) {
        log.debug("Inside the getIndexPage() method ");
        model.addAttribute("recipes", recipeService.getRecipes());
        return "indexp";
    }
}
